--- gui/actions.py
@@ -1,22 +1,0 @@
-from logging import getLogger
-from inoio import errors
-from gui.extensions import conn
-
-LOGGER = getLogger("inodaqv2")
-
-
-def run_handshake() -> None:
-    LOGGER.info("Handshaking with device")
-    LOGGER.info('Sending command: "hello"')
-
-    try:
-        conn.write("hello")
-    except errors.InoIOTransmissionError as e:
-        LOGGER.exception("Failed to send command")
-        raise ConnectionError("Could not connect to device") from e
-
-    reply = conn.read()
-
-    if reply != "1;Hello from InoDAQV2":
-        LOGGER.error('Handshake returned unknown message: "%s"', reply)
-        raise ConnectionError("Handshake returned unknown message")
